test(switchboard): one isolation test file per flag experiment
Program-isolation tests move out of switchboard.test.ts into
switchboard/flags/__tests__/, one file per experiment:
- basic-integration.test.ts — trio routing, variants, effort, cloud
  gate, and a registry-wide sweep proving the trio touches nothing else
- self-driving.test.ts — payload routing/pinning, the fail-closed
  shapes, cloud gate, and its registry-wide isolation sweep
- orchestrator.test.ts — declared-programs routing, the exact-scope
  pin (posthog-integration only), uncovered-program sweep, and the
  2026-07-17 self-driving regression test

switchboard.test.ts keeps only machinery: binding registry lockstep,
CLI precedence, trace stamping, modelCapabilities, and the composed
linear clamp.

Generated-By: PostHog Code
Task-Id: c8dcf8d0-3565-4c83-8929-3c7faecdcb1c
